requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.0
matplotlib>=3.4.0
ipywidgets>=7.6.0
jupyter>=1.0.0
streamlit>=1.28.0
